requirements.txt
azure-functions
xlrd
pandas
azure-identity
azure-storage-file-datalake
